musibara-api/requirements.txt
alabaster==0.7.16
annotated-types==0.7.0
anyio==4.6.0
astroid==3.3.5
asyncio==3.4.3
asyncpg==0.29.0
babel==2.16.0
bcrypt==4.2.0
boto3==1.35.54
botocore==1.35.54
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
default==0.1.2
dill==0.3.9
dnspython==2.6.1
docutils==0.16
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
futures==3.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
imagesize==1.4.1
isort==5.13.2
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
musicbrainzngs==0.7.1
packaging==24.2
passlib==1.7.4
platformdirs==4.3.6
psycopg==3.2.3
psycopg2==2.9.9
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pylint==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
rich==13.9.1
s3==3.0.0
s3transfer==0.10.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spotipy==2.24.0
starlette==0.38.6
tomlkit==0.13.2
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
wheel==0.44.0
xmltodict==0.14.2
